+++ example/kotlin/one-consumer-one-provider/consumer/src/main/kotlin/br/com/zup/pact/consumer/integration/account/service/AccountIntegrationService.kt
@@ -1,4 +1,4 @@
-package br.com.zup.pact.consumer.integration.service
+package br.com.zup.pact.consumer.integration.account.service
 
 import br.com.zup.pact.consumer.dto.BalanceDTO
