+++ example/kotlin/one-consumer-one-provider/consumer/src/main/kotlin/br/com/zup/pact/consumer/service/ClientServiceImpl.kt
@@ -3,7 +3,7 @@ package br.com.zup.pact.consumer.service
 import br.com.zup.pact.consumer.dto.BalanceDTO
 import br.com.zup.pact.consumer.dto.ClientDetailsDTO
 import br.com.zup.pact.consumer.exception.ClientNotFoundException
-import br.com.zup.pact.consumer.integration.service.AccountIntegrationService
+import br.com.zup.pact.consumer.integration.account.service.AccountIntegrationService
 import br.com.zup.pact.consumer.repository.ClientRepository
 import org.springframework.beans.factory.annotation.Autowired
